+++ test/vitest/__tests__/routes.test.js
@@ -1,8 +1,14 @@
-import { describe, it, expect } from 'vitest'
+import { describe, it, expect, beforeEach } from 'vitest'
 import { createRouter, createMemoryHistory } from 'vue-router'
+import { createPinia, setActivePinia } from 'pinia'
 import routes from '@/router/routes'
+import { useAuthenticateStore } from '@/stores/authenticate-store'
 
 describe('Router', () => {
+  beforeEach(() => {
+    setActivePinia(createPinia())
+  })
+
   it('should allow access to the search page without entering a neighborhood', async () => {
     const router = createRouter({
       history: createMemoryHistory(),
@@ -13,5 +19,20 @@ describe('Router', () => {
     await router.isReady()
 
     expect(router.currentRoute.value.name).toBe('search')
+  })
+
+  it('should redirect to the home page if the user is not authenticated', async () => {
+    const router = createRouter({
+      history: createMemoryHistory(),
+      routes,
+    })
+
+    router.push('/favorites')
+    await router.isReady()
+
+    const { isAuthenticated } = useAuthenticateStore()
+
+    expect(isAuthenticated).toBe(false)
+    expect(router.currentRoute.value.name).toBe('home')
   })
 })